Split classes in proper files
- Extract IPrintable from main.cpp
- Extract Coordinate from main.cpp
- Extract Cell from main.cpp
- Extract Grid from main.cpp
