+++ iprintable.h
@@ -1,0 +1,10 @@
+#ifndef IPRINTABLE_H
+#define IPRINTABLE_H
+
+class IPrintable
+{
+public:
+    virtual void print() const = 0;
+};
+
+#endif // IPRINTABLE_H
